# internal/port/http/docs/swagger.yaml
definitions:
  model.Animal:
    properties:
      animalTypeGuid:
        type: string
      descriptive:
        type: string
      isFed:
        type: boolean
      isIll:
        type: boolean
      weight:
        type: number
    type: object
info:
  contact: {}
paths:
  /animals:
    get:
      description: Get a list of all animals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Animal'
            type: array
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve a list of all animals
      tags:
      - animals
    post:
      consumes:
      - application/json
      description: Create a new animal with the provided details
      parameters:
      - description: Animal
        in: body
        name: animal
        required: true
        schema:
          $ref: '#/definitions/model.Animal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Animal'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new animal
      tags:
      - animals
    put:
      consumes:
      - application/json
      description: Update an animal with the provided details
      parameters:
      - description: Animal
        in: body
        name: animal
        required: true
        schema:
          $ref: '#/definitions/model.Animal'
      produces:
      - application/json
      responses:
        "200":
          description: updated
          schema:
            type: string
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing animal
      tags:
      - animals
  /animals/{guid}:
    delete:
      description: Delete an animal by its GUID
      parameters:
      - description: GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an animal by its GUID
      tags:
      - animals
    get:
      description: Get details of an animal by its GUID
      parameters:
      - description: GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Animal'
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve an animal by its GUID
      tags:
      - animals
swagger: "2.0"